refactor: Move limitbuffer -> sourcenode/limitbuffer
